tools/requirements.txt
mitmproxy>=9.0.0,<=10.2.0
dnspython
httpx[socks]
packaging
Pillow